fix: fully specify imports (#44)
In this package, we have a mix of fully specified (including .js extension) in our index.js
and non-fully specified in two instances of pages. With the cypress v12 upgrade this package
is expected to have only fully resolved imports. We can change this in the future with a
custom webpack preprocessor in our cypress e2e.js file, but that seemed a bit overkill to me
for these two import instances.

Signed-off-by: Daniel Rowe <rowdane@amazon.com>
Co-authored-by: Daniel Rowe <rowdane@amazon.com>